+++ install_defaults.php
@@ -1,33 +1,4 @@
 <?php
-// +--------------------------------------------------------------------------+
-// | evList Plugin - glFusion CMS                                             |
-// +--------------------------------------------------------------------------+
-// | install_defaults.php                                                     |
-// |                                                                          |
-// | Initial Installation Defaults used when loading the online configuration |
-// | records. These settings are only used during the initial installation    |
-// | and not referenced any more once the plugin is installed.                |
-// +--------------------------------------------------------------------------+
-// | Copyright (C) 2008-2010 by the following authors:                        |
-// |                                                                          |
-// | Mark R. Evans          mark AT glfusion DOT org                          |
-// +--------------------------------------------------------------------------+
-// |                                                                          |
-// | This program is free software; you can redistribute it and/or            |
-// | modify it under the terms of the GNU General Public License              |
-// | as published by the Free Software Foundation; either version 2           |
-// | of the License, or (at your option) any later version.                   |
-// |                                                                          |
-// | This program is distributed in the hope that it will be useful,          |
-// | but WITHOUT ANY WARRANTY; without even the implied warranty of           |
-// | MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the            |
-// | GNU General Public License for more details.                             |
-// |                                                                          |
-// | You should have received a copy of the GNU General Public License        |
-// | along with this program; if not, write to the Free Software Foundation,  |
-// | Inc., 59 Temple Place - Suite 330, Boston, MA  02111-1307, USA.          |
-// |                                                                          |
-// +--------------------------------------------------------------------------+
 /**
 *   Installation defaults for the evList plugin
 *
@@ -359,6 +330,17 @@ $evlistConfigData = array(
         'fieldset' => 20,
         'selection_array' => 10,
         'sort' => 70,
+        'set' => true,
+        'group' => 'evlist',
+    ),
+    array(
+        'name' => 'cb_hide_small',
+        'default_value' => 0,
+        'type' => 'select',
+        'subgroup' => 0,
+        'fieldset' => 20,
+        'selection_array' => 0,
+        'sort' => 80,
         'set' => true,
         'group' => 'evlist',
     ),